==> R/02_clean.qmd <==
---
title: "Clean"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}
library(ggplot2)
library(tidyverse)
```

```{r}
data <- read.csv("../data/loaded.csv")
```

```{r}
# Drop NA and remove 
cleaned_data <- data |> drop_na() |> select(-row.names)
```

```{r}
cat("Number of rows dropped:", nrow(data) - nrow(cleaned_data),"\n")
```

```{r}
cat("Number of colums dropped:", ncol(data) - ncol(cleaned_data),"\n")
```

```{r}
write_csv(x = cleaned_data, file = "../data/cleaned.csv")
```
