==> R/05_analysis.qmd <==
---
title: "Analysis"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}
library(tidyverse)
library(ggplot2)
```

Import data

```{r}
data <- read.csv("../data/augmented.csv")
```

Tobacco and alcohol

```{r}
col_colours <- c("0" = "#0053B3", 
                 "1"="#FFAA17")


exposure_behaviour <- ggplot(data=data,
                             mapping = aes(x=alcohol,
                                           y=tobacco,
                                           color=factor(chd))) +
    scale_color_manual(values = col_colours) +
  geom_point() +
  labs(title = "Behavioural exposures (tobacco and alcohol) and coronary heart disease", 
       x = "Alcohol", 
       y = "Tobacco", 
       color  = "Coronary heart disease")

exposure_behaviour

```

Obesity and adiposity

```{r}

indicators_meta <- ggplot(data=data,
                          mapping = aes(x=adiposity,
                                        y=obesity,
                                        color=factor(chd))) +
  geom_point() + 
  geom_smooth(method = 'loess', 
              formula = 'y~x') + 
  scale_color_manual(values = col_colours) +

  labs(title = "Metabolic indicators (adiposity and obesity) and coronary heart disease", 

       x = "Adiposity", 
       y = "Obesity", 
       color  = "Coronary heart disease")

indicators_meta

```

Investigating family history

```{r}

fam_history <- ggplot(data,
                      mapping = aes(x=inheritance)) +
  geom_bar() +
  theme_minimal() +
  labs(title = "Cardiovascular disease by family history", x = "Inheritance pattern", y = "Count")

fam_history


```

```{r}

ggsave(filename = "05a_Exposure.png", plot = exposure_behaviour, path = "../results/")
ggsave(filename = "05b_Metabolic indicators.png", plot = indicators_meta, path = "../results/")
ggsave(filename = "05c_Family history.png", plot = fam_history, path = "../results/")

```
